requirements.txt
crewai
crewai-tools
openai
gradio
load_dotenv


